requirements.txt
google-genai==1.26.0
streamlit==1.48.1
streamlit-chat==0.1.1
jsonpickle==4.0.5
python-dotenv==1.0.1
replicate==1.0.4
chromadb==0.4.22
numpy<2.0
openai==1.101.0
pymongo==4.15.2